requirements.txt
CairoSVG==1.0.20
Flask==0.10.1
Flask-Login==0.2.11
Flask-Migrate==1.6.0
Flask-SQLAlchemy==2.0
Flask-Script==2.0.5
Jinja2==2.7.3
Mako==1.0.2
MarkupSafe==0.23
MySQL-python==1.2.5
Pillow==3.0.0
Pyphen==0.9.2
SQLAlchemy==1.0.5
WeasyPrint==0.24
Werkzeug==0.10.4
alembic==0.8.2
argparse==1.2.1
bcrypt==1.1.1
cairocffi==0.7.2
cffi==1.1.1
cssselect==0.9.1
dateutils==0.6.6
google-api-python-client==1.5.0
greenlet==0.4.9
html5lib==0.9999999
httplib2==0.9.2
itsdangerous==0.24
lxml==3.5.0
mysql==0.0.1
oauth2client==2.0.1
psycopg2==2.6
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.13
python-dateutil==2.5.0
python-editor==0.4
pytz==2015.7
requests==2.6.0
rsa==3.4.2
simplejson==3.8.2
smtpapi==0.2.0
uWSGI==2.0.12
uritemplate==0.6
wheel==0.26.0
wsgiref==0.1.2
